grove_rotary_angle_sensor.py
#!/usr/bin/python3

# -*- coding: utf-8 -*-
#
# The MIT License (MIT)
#
# Grove Base Hat for the Raspberry Pi, used to connect grove sensors.
# Copyright (C) 2018  Seeed Technology Co.,Ltd.

from grove.adc import ADC

__all__ = ["GroveRotaryAngleSensor"]

class GroveRotaryAngleSensor(ADC):
    '''
    Grove Rotary Angle Sensor class

    Args:
        pin(int): number of analog pin/channel the sensor connected.
    '''
    def __init__(self, channel):
        self.channel = channel
        self.adc = ADC()
    
    @property
    def value(self):
        '''
        Get the rotary angle value, max angle is 100.0%

        Returns:
            (int): ratio, 0(0.0%) - 1000(100.0%)
        '''
        return self.adc.read(self.channel)

