requirements.txt
RPi.GPIO
pigpio
grove.py
paho-mqtt
selenium